fix hive2 recurrence find peer limit

+++ pkg/hive2/hive2.go
@@ -158,7 +158,7 @@ func (s *Service) onFindNode(ctx context.Context, peer p2p.Peer, stream p2p.Stre
 	return nil
 }
 
-func (s *Service) DoFindNode(ctx context.Context, peer boson.Address, pos []int32, limit int32) (res chan boson.Address, total int, err error) {
+func (s *Service) DoFindNode(ctx context.Context, peer boson.Address, pos []int32, limit int32) (res chan boson.Address, total int32, err error) {
 	s.metrics.DoFindNode.Inc()
 	stream, err := s.streamer.NewStream(ctx, peer, nil, protocolName, protocolVersion, streamFindNode)
 	if err != nil {
@@ -199,7 +199,7 @@ func (s *Service) DoFindNode(ctx context.Context, peer boson.Address, pos []int3
 		return res, 0, errors.New("failed to process peers, shutting down hive2")
 	}
 
-	return res, len(result.Peers), nil
+	return res, int32(len(result.Peers)), nil
 }
 
 func (s *Service) startCheckPeersHandler() {
